pyproject.toml
[tool.poetry]
name = "talentai-pro"
version = "0.1.0"
description = "AI-Powered Talent Intelligence Platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"

# FastAPI & Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.18"

# Database & ORM
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
asyncpg = "^0.30.0"
pymongo = "^4.15.0"
redis = "^5.2.0"

# Validation & Configuration
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# ML & AI Libraries
torch = "^2.5.0"
transformers = "^4.46.0"
sentence-transformers = "^3.2.0"
faiss-cpu = "^1.9.0"
nltk = "^3.9.1"
numpy = "^2.1.0"

# Background Jobs
celery = "^5.4.0"

# HTTP Client
httpx = "^0.28.0"
email-validator = "^2.3.0"
pyjwt = "^2.10.1"
argon2-cffi = "^25.1.0"
pytest-asyncio = "^1.2.0"
aiosqlite = "^0.21.0"
aioredis = "^2.0.1"
elasticsearch = "^9.2.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
scikit-learn = "^1.7.2"
hiredis = "^3.3.0"
python-docx = "^1.2.0"
pypdf2 = "^3.0.1"
pypdf = "^6.1.3"
pytesseract = "^0.3.13"
pillow = "^12.0.0"
spacy = "^3.8.8"
pandas = "^2.3.3"
plotly = "^6.4.0"
mlflow = "^3.6.0"
wandb = "^0.22.3"
langchain = "^1.0.5"
langchain-openai = "^1.0.2"
langchain-anthropic = "^1.1.0"
langgraph = "^1.0.3"
chromadb = "^1.3.4"
ray = {version = "^2.51.1", extras = ["default"]}
openai = "^2.7.1"
requests = "^2.32.5"
psutil = "^7.1.3"
psycopg2-binary = "^2.9.11"
python-socketio = "^5.14.3"
status = "^0.2.5"
redis-om = "^0.3.5"
google-generativeai = "^0.8.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
